+++ NEWS.md
@@ -1,3 +1,10 @@
+# platjam version 0.0.2.900
+
+## new function
+
+* `nmatlist2heatmaps()` creates multiple coverage heatmaps
+as the result of importing data with `coverage_matrix2nmat()`.
+
 # platjam version 0.0.1.900
 
 Initial release.